Corrected compilation with MSVC 11.
git-svn-id: https://boost-log.svn.sourceforge.net/svnroot/boost-log/branches/v1@801 ab6a8de6-4e00-4c3d-a860-8899c17d353f

+++ boost/log/utility/string_literal.hpp
@@ -496,7 +496,7 @@ private:
             if (result != 0)
                 return result;
         }
-        return (LeftLen - RightLen);
+        return static_cast< int >(LeftLen - RightLen);
     }
 #endif // BOOST_LOG_DOXYGEN_PASS
 };
